+++ src/models/json/museum-json-store.js
@@ -33,6 +33,7 @@ export const museumJsonStore = {
   async deleteMuseumById(id) {
     await db.read();
     const index = db.data.museums.findIndex((m) => m._id === id);
+    await exhibitionJsonStore.deleteExhibitionsByMuseumId(museum._id);
     db.data.museums.splice(index, 1);
     await db.write();
   },